labs/04-pdm-cluster-k8s/inference/requirements.txt
paho-mqtt
joblib
scikit-learn
pandas
numpy